MySQL - MySQL - rudimentary implementations to pass all Test_1_Mutation tests

+++ Json/Fliox.Hub.MySQL/MySQLContainer.cs
@@ -4,10 +4,12 @@
 #if !UNITY_5_3_OR_NEWER || MYSQL
 
 using System;
+using System.Text;
 using System.Threading.Tasks;
 using Friflo.Json.Fliox.Hub.Host;
 using Friflo.Json.Fliox.Hub.Protocol.Models;
 using Friflo.Json.Fliox.Hub.Protocol.Tasks;
+using MySqlConnector;
 
 
 namespace Friflo.Json.Fliox.Hub.MySQL
@@ -29,20 +31,44 @@ namespace Friflo.Json.Fliox.Hub.MySQL
             if (tableExists) {
                 return null;
             }
-            var sql = $"CREATE TABLE if not exists {name} (id VARCHAR(128), data TEXT);";
-            await MySQLUtils.Execute(database.connection, sql).ConfigureAwait(false);
+            var sql = $"CREATE TABLE if not exists {name} (id VARCHAR(128) PRIMARY KEY, data TEXT);";
+            var result = await MySQLUtils.Execute(database.connection, sql).ConfigureAwait(false);
+            if (result.error != null) {
+                return result.error;
+            }
             tableExists = true;
             return null;
         }
         
-        public override Task<CreateEntitiesResult> CreateEntitiesAsync(CreateEntities command, SyncContext syncContext) {
-            var result = CreateEntities(command, syncContext);
-            return Task.FromResult(result);
+        public override async Task<CreateEntitiesResult> CreateEntitiesAsync(CreateEntities command, SyncContext syncContext) {
+            var error = await EnsureContainerExists().ConfigureAwait(false);
+            if (error != null) {
+                return new CreateEntitiesResult { Error = error };
+            }
+            var sql = new StringBuilder();
+            sql.Append($"INSERT INTO {name} (id,data) VALUES\n");
+            MySQLUtils.AppendValues(sql, command.entities);
+            using var cmd = new MySqlCommand(sql.ToString(), database.connection);
+            try {
+                await cmd.ExecuteNonQueryAsync().ConfigureAwait(false);
+            } catch (Exception e) {
+                return new CreateEntitiesResult { Error = DatabaseError(e.Message) };    
+            }
+            return new CreateEntitiesResult();
         }
         
-        public override Task<UpsertEntitiesResult> UpsertEntitiesAsync(UpsertEntities command, SyncContext syncContext) {
-            var result = UpsertEntities(command, syncContext);
-            return Task.FromResult(result);
+        public override async Task<UpsertEntitiesResult> UpsertEntitiesAsync(UpsertEntities command, SyncContext syncContext) {
+            var error = await EnsureContainerExists().ConfigureAwait(false);
+            if (error != null) {
+                return new UpsertEntitiesResult { Error = error };
+            }
+            var sql = new StringBuilder();
+            sql.Append($"REPLACE INTO {name} (id,data) VALUES\n");
+            MySQLUtils.AppendValues(sql, command.entities);
+            using var cmd = new MySqlCommand(sql.ToString(), database.connection);
+            await cmd.ExecuteNonQueryAsync().ConfigureAwait(false);
+
+            return new UpsertEntitiesResult();
         }
 
         public override Task<ReadEntitiesResult> ReadEntitiesAsync(ReadEntities command, SyncContext syncContext) {
@@ -69,13 +95,27 @@ namespace Friflo.Json.Fliox.Hub.MySQL
         
        
         public override async Task<DeleteEntitiesResult> DeleteEntitiesAsync(DeleteEntities command, SyncContext syncContext) {
-            await EnsureContainerExists().ConfigureAwait(false);
+            var error = await EnsureContainerExists().ConfigureAwait(false);
+            if (error != null) {
+                return new DeleteEntitiesResult { Error = error };
+            }
             if (command.all == true) {
                 var sql = $"DELETE from {name}";
                 await MySQLUtils.Execute(database.connection, sql).ConfigureAwait(false);
                 return new DeleteEntitiesResult();    
+            } else {
+                var sql = new StringBuilder();
+                sql.Append($"DELETE FROM  {name} WHERE id in\n");
+                
+                MySQLUtils.AppendKeys(sql, command.ids);
+                using var cmd = new MySqlCommand(sql.ToString(), database.connection);
+                await cmd.ExecuteNonQueryAsync().ConfigureAwait(false);
+                return new DeleteEntitiesResult();
             }
-            return new DeleteEntitiesResult();
+        }
+        
+        private static TaskExecuteError DatabaseError(string message) {
+            return new TaskExecuteError(TaskErrorType.DatabaseError, message);
         }
     }
 }
